+++ tests/unit/test_friends.py
@@ -27,8 +27,8 @@ class TestFriends(tests.unit.TestResource):
             self.app_secret,
             app_redirect_uri=self.app_redirect_uri,
             app_scope=self.app_scope,
-            user_token=self.token)
-        mock_get.assert_called_with(friends, upapi.endpoints.FRIENDS)
+            user_credentials=self.credentials)
+        mock_get.assert_called_with(friends, upapi.endpoints.USERFRIENDS)
         for index, friend in enumerate(friends.items):
             self.assertEqual(friend.xid, friends_data['items'][index]['xid'])
         self.assertEqual(friends.size, friends_data['size'])