+++ upapi/exceptions.py
@@ -8,3 +8,10 @@ class UnexpectedAPIResponse(Exception):
     UpApi raises this for API responses other than expected (according to the documentation)
     """
     pass
+
+
+class MissingCredentials(Exception):
+    """
+    Raised when trying to act on behalf of the user without setting the Credentials object.
+    """
+    pass